+++ ui/src/timeMachine/components/Queries.tsx
@@ -7,7 +7,9 @@ import TimeMachineFluxEditor from 'src/timeMachine/components/TimeMachineFluxEdi
 import CSVExportButton from 'src/shared/components/CSVExportButton'
 import TimeMachineQueriesSwitcher from 'src/timeMachine/components/QueriesSwitcher'
 import TimeMachineRefreshDropdown from 'src/timeMachine/components/RefreshDropdown'
-import TimeRangeDropdown from 'src/shared/components/TimeRangeDropdown'
+import TimeRangeDropdown, {
+  RangeType,
+} from 'src/shared/components/TimeRangeDropdown'
 import TimeMachineQueryTab from 'src/timeMachine/components/QueryTab'
 import TimeMachineQueryBuilder from 'src/timeMachine/components/QueryBuilder'
 import SubmitQueryButton from 'src/timeMachine/components/SubmitQueryButton'
@@ -24,32 +26,41 @@ import {
 } from '@influxdata/clockface'
 
 // Actions
-import {addQuery} from 'src/timeMachine/actions'
+import {addQuery, setAutoRefresh} from 'src/timeMachine/actions'
 import {setTimeRange} from 'src/timeMachine/actions'
 
 // Utils
 import {getActiveTimeMachine, getActiveQuery} from 'src/timeMachine/selectors'
 
 // Types
-import {AppState, DashboardQuery, QueryEditMode, TimeRange} from 'src/types'
+import {
+  AppState,
+  DashboardQuery,
+  QueryEditMode,
+  TimeRange,
+  AutoRefresh,
+  AutoRefreshStatus,
+} from 'src/types'
 import {DashboardDraftQuery} from 'src/types/dashboards'
 
 interface StateProps {
   activeQuery: DashboardQuery
   draftQueries: DashboardDraftQuery[]
   timeRange: TimeRange
+  autoRefresh: AutoRefresh
 }
 
 interface DispatchProps {
   onAddQuery: typeof addQuery
   onSetTimeRange: typeof setTimeRange
+  onSetAutoRefresh: typeof setAutoRefresh
 }
 
 type Props = StateProps & DispatchProps
 
 class TimeMachineQueries extends PureComponent<Props> {
   public render() {
-    const {draftQueries, onAddQuery, timeRange, onSetTimeRange} = this.props
+    const {draftQueries, onAddQuery, timeRange} = this.props
 
     return (
       <div className="time-machine-queries">
@@ -82,7 +93,7 @@ class TimeMachineQueries extends PureComponent<Props> {
               <TimeMachineRefreshDropdown />
               <TimeRangeDropdown
                 timeRange={timeRange}
-                onSetTimeRange={onSetTimeRange}
+                onSetTimeRange={this.handleSetTimeRange}
               />
               <TimeMachineQueriesSwitcher />
               <SubmitQueryButton />
@@ -92,6 +103,29 @@ class TimeMachineQueries extends PureComponent<Props> {
         <div className="time-machine-queries--body">{this.queryEditor}</div>
       </div>
     )
+  }
+
+  private handleSetTimeRange = (
+    timeRange: TimeRange,
+    rangeType: RangeType = RangeType.Relative
+  ) => {
+    const {autoRefresh, onSetAutoRefresh, onSetTimeRange} = this.props
+
+    onSetTimeRange(timeRange)
+
+    if (rangeType === RangeType.Absolute) {
+      onSetAutoRefresh({...autoRefresh, status: AutoRefreshStatus.Disabled})
+      return
+    }
+
+    if (autoRefresh.status === AutoRefreshStatus.Disabled) {
+      if (autoRefresh.interval === 0) {
+        onSetAutoRefresh({...autoRefresh, status: AutoRefreshStatus.Paused})
+        return
+      }
+
+      onSetAutoRefresh({...autoRefresh, status: AutoRefreshStatus.Active})
+    }
   }
 
   private get queryEditor(): JSX.Element {
@@ -108,16 +142,17 @@ class TimeMachineQueries extends PureComponent<Props> {
 }
 
 const mstp = (state: AppState) => {
-  const {draftQueries, timeRange} = getActiveTimeMachine(state)
+  const {draftQueries, timeRange, autoRefresh} = getActiveTimeMachine(state)
 
   const activeQuery = getActiveQuery(state)
 
-  return {timeRange, activeQuery, draftQueries}
+  return {timeRange, activeQuery, draftQueries, autoRefresh}
 }
 
 const mdtp = {
   onAddQuery: addQuery,
   onSetTimeRange: setTimeRange,
+  onSetAutoRefresh: setAutoRefresh,
 }
 
 export default connect<StateProps, DispatchProps>(